+++ database/factories/ModelFactory.php
@@ -20,7 +20,6 @@ $factory->define(Page::class, function (Faker\Generator $faker) {
     return [
         'morph_key' => 'singles',
         'published'   => 1,
-        'featured'    => $faker->boolean(),
         'publication' => null,
         'title:nl'    => $faker->words(rand(2, 4), true),
     ];