+++ src/Urls/UrlHelper.php
@@ -2,6 +2,9 @@
 
 namespace Thinktomorrow\Chief\Urls;
 
+use Illuminate\Support\Facades\DB;
+use Illuminate\Database\Eloquent\Model;
+use Thinktomorrow\Chief\FlatReferences\FlatReference;
 use Thinktomorrow\Chief\Concerns\Morphable\Morphables;
 use Thinktomorrow\Chief\FlatReferences\FlatReferencePresenter;
 
@@ -11,15 +14,27 @@ class UrlHelper
      * Internal api for fetching all models that have an active url
      *
      * @param bool $onlySingles
+     * @param Model|null $ignoredModel
      * @return array
      */
-    public static function allOnlineModels(bool $onlySingles = false): array
+    public static function allOnlineModels(bool $onlySingles = false, Model $ignoredModel = null): array
     {
-        return chiefMemoize('all-online-models', function () use ($onlySingles) {
-            $builder = UrlRecord::whereNull('redirect_id')->select('model_type', 'model_id')->groupBy('model_type', 'model_id');
+        return chiefMemoize('all-online-models', function () use ($onlySingles, $ignoredModel) {
+            $builder = UrlRecord::whereNull('redirect_id')
+                ->select('model_type', 'model_id')
+                ->groupBy('model_type', 'model_id');
 
             if ($onlySingles) {
                 $builder->where('model_type', 'singles');
+            }
+
+            if ($ignoredModel) {
+                $builder->whereNotIn('id', function ($query) use ($ignoredModel) {
+                    $query->select('id')
+                        ->from('chief_urls')
+                        ->where('model_type', '=', $ignoredModel->getMorphClass())
+                        ->where('model_id', '=', $ignoredModel->id);
+                });
             }
 
             $liveUrlRecords = $builder->get()->mapToGroups(function ($record) {
@@ -35,7 +50,7 @@ class UrlHelper
             })->flatten();
 
             return FlatReferencePresenter::toGroupedSelectValues($models)->toArray();
-        });
+        }, [$onlySingles, $ignoredModel]);
     }
 
     public static function allOnlineSingles()